requirements.txt
certifi==2020.4.5.1
Django==2.2.9
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-js-asset==1.2.2
Pillow==7.1.2
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
